Move all of this shipdata stuff into a new package
