+++ src/main/java/org/valkyrienskies/mod/common/command/MainCommand.java
@@ -7,8 +7,8 @@ import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.world.World;
 import org.valkyrienskies.mod.common.multithreaded.VWThread;
-import org.valkyrienskies.mod.common.physmanagement.interaction.QueryableShipData;
-import org.valkyrienskies.mod.common.physmanagement.interaction.ShipData;
+import org.valkyrienskies.mod.common.physmanagement.shipdata.QueryableShipData;
+import org.valkyrienskies.mod.common.physmanagement.shipdata.ShipData;
 import org.valkyrienskies.mod.common.ship_handling.IHasShipManager;
 import org.valkyrienskies.mod.common.ship_handling.WorldServerShipManager;
 import org.valkyrienskies.mod.common.util.ValkyrienUtils;
