+++ src/main/java/org/valkyrienskies/mod/common/physmanagement/shipdata/IValkyrienSkiesWorldData.java
@@ -1,7 +1,7 @@
-package org.valkyrienskies.mod.common.physmanagement.chunk;
+package org.valkyrienskies.mod.common.physmanagement.shipdata;
 
 import net.minecraft.nbt.NBTTagCompound;
-import org.valkyrienskies.mod.common.physmanagement.interaction.QueryableShipData;
+import org.valkyrienskies.mod.common.physmanagement.chunk.ShipChunkAllocator;
 
 public interface IValkyrienSkiesWorldData {
 
@@ -12,4 +12,5 @@ public interface IValkyrienSkiesWorldData {
     ShipChunkAllocator getChunkAllocator();
 
     QueryableShipData getQueryableShipData();
+
 }
